Some windows build fixes. MPIR-2.7.0 -> MPIR-2.7.2.

+++ deps/build.jl
@@ -19,13 +19,21 @@ DLCFLAGS = "-fPIC -fno-common"
 
 cd(wdir)
 
+function download_dll(url_string, location_string)
+   try
+      run(`curl -o $(location_string) -L $(url_string)`)
+   catch
+      download(url_string, location_string)
+   end
+end
+
 #install libpthreads
 
 if on_windows
    if Int == Int32
-      download("http://nemocas.org/binaries/w32-libwinpthread-1.dll", joinpath(vdir, "lib", "libwinpthread-1.dll"))
+      download_dll("http://nemocas.org/binaries/w32-libwinpthread-1.dll", joinpath(vdir, "lib", "libwinpthread-1.dll"))
    else
-      download("http://nemocas.org/binaries/w64-libwinpthread-1.dll", joinpath(vdir, "lib", "libwinpthread-1.dll"))
+      download_dll("http://nemocas.org/binaries/w64-libwinpthread-1.dll", joinpath(vdir, "lib", "libwinpthread-1.dll"))
    end
 end
 
@@ -33,34 +41,36 @@ cd(wdir)
 
 # install M4
 
-try
-   run(`m4 --version`)
-catch
-   download("http://ftp.gnu.org/gnu/m4/m4-1.4.17.tar.bz2", joinpath(wdir, "m4-1.4.17.tar.bz2"))
-   run(`tar -xvf m4-1.4.17.tar.bz2`)
-   run(`rm m4-1.4.17.tar.bz2`)
-   cd("$wdir/m4-1.4.17")
-   run(`./configure --prefix=$vdir`)
-   run(`make`)
-   run(`make install`)
+if !on_windows
+   try
+      run(`m4 --version`)
+   catch
+      download("http://ftp.gnu.org/gnu/m4/m4-1.4.17.tar.bz2", joinpath(wdir, "m4-1.4.17.tar.bz2"))
+      run(`tar -xvf m4-1.4.17.tar.bz2`)
+      run(`rm m4-1.4.17.tar.bz2`)
+      cd("$wdir/m4-1.4.17")
+      run(`./configure --prefix=$vdir`)
+      run(`make`)
+      run(`make install`)
+   end
 end
 
 cd(wdir)
 
 # install GMP/MPIR
 
-if !ispath(Pkg.dir("Nemo", "local", "mpir-2.7.0"))
-   download("http://mpir.org/mpir-2.7.0.tar.bz2", joinpath(wdir, "mpir-2.7.0.tar.bz2"))
-   run(`tar -xvf mpir-2.7.0.tar.bz2`)
-   run(`rm mpir-2.7.0.tar.bz2`)
+if !ispath(Pkg.dir("Nemo", "local", "mpir-2.7.2"))
+   download("http://mpir.org/mpir-2.7.2.tar.bz2", joinpath(wdir, "mpir-2.7.2.tar.bz2"))
+   run(`tar -xvf mpir-2.7.2.tar.bz2`)
+   run(`rm mpir-2.7.2.tar.bz2`)
 end
-cd("$wdir/mpir-2.7.0")
+cd("$wdir/mpir-2.7.2")
 
 if on_windows
    if Int == Int32
-      download("http://nemocas.org/binaries/w32-libgmp-10.dll", joinpath(vdir, "lib", "libgmp-10.dll"))
+      download_dll("http://nemocas.org/binaries/w32-libgmp-16.dll", joinpath(vdir, "lib", "libgmp-16.dll"))
    else
-      download("http://nemocas.org/binaries/w64-libgmp-10.dll", joinpath(vdir, "lib", "libgmp-10.dll"))
+      download_dll("http://nemocas.org/binaries/w64-libgmp-16.dll", joinpath(vdir, "lib", "libgmp-16.dll"))
    end
 else
    try
@@ -87,9 +97,9 @@ end
 
 if on_windows
    if Int == Int32
-      download("http://nemocas.org/binaries/w32-libmpfr-4.dll", joinpath(vdir, "lib", "libmpfr-4.dll"))
+      download_dll("http://nemocas.org/binaries/w32-libmpfr-4.dll", joinpath(vdir, "lib", "libmpfr-4.dll"))
    else
-      download("http://nemocas.org/binaries/w64-libmpfr-4.dll", joinpath(vdir, "lib", "libmpfr-4.dll"))
+      download_dll("http://nemocas.org/binaries/w64-libmpfr-4.dll", joinpath(vdir, "lib", "libmpfr-4.dll"))
    end
 else
    cd("$wdir/mpfr-3.1.3")
@@ -124,9 +134,9 @@ end
 
 if on_windows
    if Int == Int32
-      download("http://nemocas.org/binaries/w32-libflint.dll", joinpath(vdir, "lib", "libflint.dll"))
+      download_dll("http://nemocas.org/binaries/w32-libflint.dll", joinpath(vdir, "lib", "libflint.dll"))
    else
-      download("http://nemocas.org/binaries/w64-libflint.dll", joinpath(vdir, "lib", "libflint.dll"))
+      download_dll("http://nemocas.org/binaries/w64-libflint.dll", joinpath(vdir, "lib", "libflint.dll"))
    end
    try
       run(`ln -sf $vdir\\lib\\libflint.dll $vdir\\lib\\libflint-13.dll`)
@@ -154,9 +164,9 @@ end
  
 if on_windows
    if Int == Int32
-      download("http://nemocas.org/binaries/w32-libarb.dll", joinpath(vdir, "lib", "libarb.dll"))
+      download_dll("http://nemocas.org/binaries/w32-libarb.dll", joinpath(vdir, "lib", "libarb.dll"))
    else
-      download("http://nemocas.org/binaries/w64-libarb.dll", joinpath(vdir, "lib", "libarb.dll"))
+      download_dll("http://nemocas.org/binaries/w64-libarb.dll", joinpath(vdir, "lib", "libarb.dll"))
    end
 else
    cd("$wdir/arb")
@@ -183,9 +193,9 @@ end
 
 if on_windows
    if Int == Int32
-      download("http://nemocas.org/binaries/w32-libpari.dll", joinpath(vdir, "lib", "libpari.dll"))
+      download_dll("http://nemocas.org/binaries/w32-libpari.dll", joinpath(vdir, "lib", "libpari.dll"))
    else
-      download("http://nemocas.org/binaries/w64-libpari.dll", joinpath(vdir, "lib", "libpari.dll"))
+      download_dll("http://nemocas.org/binaries/w64-libpari.dll", joinpath(vdir, "lib", "libpari.dll"))
    end
 elseif on_osx
    cd("$wdir/pari-2.7.4")
